backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
flask-jwt-extended
python-dotenv
pymysql
cryptography
werkzeug
gunicorn
Pillow>=10.0.0
ruff
mypy
pytest
python-slugify